+++ action_server/docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `action-server package metadata` now includes information on the secrets required
   for each action.
+- Passing secrets now works (requires `robocorp-actions=0.2.0`)
 
 ## 0.2.1 - 2024-04-04